+++ src/tests/customMutation.spec.js
@@ -49,4 +49,71 @@ describe('Test the custom mutations', () => {
     )
     expect(trace).toMatchSnapshot()
   })
+
+  it('Check that you can call an additional mutation on a model', async () => {
+    const response = await request(server)
+      .post('/graphql')
+      .set('userid', 1)
+      .send({
+        query: `mutation anonymizeLog($logId: String) {
+              anonymizeLog(logId: $logId) {
+                id
+                message
+              }
+            }`,
+        variables: {
+          logId: '123',
+        },
+        operationName: 'anonymizeLog',
+      })
+    expect(response.body.errors).toBeUndefined()
+    expect(response.body.data.anonymizeLog).not.toBeUndefined()
+    expect(response.body.data.anonymizeLog).toMatchSnapshot(
+      'The anonymized log'
+    )
+    expect(trace).toMatchSnapshot()
+  })
+
+  it('Check that you can use a completely custom create mutation', async () => {
+    const response = await request(server)
+      .post('/graphql')
+      .set('userid', 1)
+      .send({
+        query: `mutation companyCreate($company: companyInput!, $setupOptions: CompanySetupOptions) {
+              companyCreate(company: $company, setupOptions: $setupOptions) {
+                id
+                name
+                users {
+                  id
+                  name
+                }
+                departments {
+                  id
+                  name
+                }
+              }
+            }`,
+        variables: {
+          company: {
+            name: 'New Company with Setup',
+            companyTypeId: 1,
+            managerId: 1,
+            type: {
+              id: 1,
+            },
+          },
+          setupOptions: {
+            createDefaultDepartment: true,
+            addDefaultUser: true,
+          },
+        },
+        operationName: 'companyCreate',
+      })
+    expect(response.body.errors).toBeUndefined()
+    expect(response.body.data.companyCreate).not.toBeUndefined()
+    expect(response.body.data.companyCreate).toMatchSnapshot(
+      'The created company with setup'
+    )
+    expect(trace).toMatchSnapshot()
+  })
 })