+++ src/types/types.d.ts
@@ -14,9 +14,12 @@ import {
   Filterable,
   FindOptions,
   GroupedCountResultItem,
+  Includeable,
+  InferAttributes,
   Model,
   ModelStatic,
   Sequelize,
+  WhereAttributeHash,
 } from 'sequelize/types'
 
 export type Action = 'list' | 'create' | 'delete' | 'update' | 'count'
@@ -31,7 +34,6 @@ export type ExtraArg = { [key: string]: { type: GraphQLType } }
 
 export type TSource = any
 export type TArgs = { [key: string]: any }
-export type TContext = any
 export type TInfo = any
 
 export type Payload = any
@@ -62,76 +64,81 @@ export type EndpointArgs = {
   [key: string]: EnpointArg
 }
 
-export type CustomResolver = (
+export type CustomResolver<TContext = any> = (
   source: any,
   args: TArgs,
   context: TContext,
   info: TInfo
 ) => Promise<any>
 
-export type CustomMutationConfiguration = {
+export type CustomMutationConfiguration<TContext = any> = {
   type: GraphQLObjectType | GraphQLList<GraphQLObjectType>
   description?: string
   args: EndpointArgs
-  resolve: CustomResolver
+  resolve: CustomResolver<TContext>
 }
 
-export type MutationList = {
-  [key: string]: CustomMutationConfiguration
+export type MutationList<TContext = any> = {
+  [key: string]: CustomMutationConfiguration<TContext>
 }
 
-export type CustomSubscriptionConfiguration = {
+export type CustomSubscriptionConfiguration<TContext = any> = {
   type: GraphQLObjectType | GraphQLList<GraphQLObjectType>
   description?: string
   args: EndpointArgs
   subscribe: GraphQLFieldResolver<TSource, TContext, TArgs>
 }
 
-export type SubscriptionList = {
-  [key: string]: CustomSubscriptionConfiguration
+export type SubscriptionList<TContext = any> = {
+  [key: string]: CustomSubscriptionConfiguration<TContext>
 }
 
 export type setWebhookDataType = (defaultData: any) => (f: Function) => void
 
-export type GlobalBeforeHook = (params: {
+export type GlobalBeforeHook<TContext = any> = (params: {
   args: TArgs
   context: TContext
   info: TInfo
 }) => any | Promise<any>
 
-export type QueryBeforeHook<M extends Model<any>> = (params: {
-  findOptions: FindOptions<M>
-  args: any
-  context: any
-  info: any
-}) => Promise<FindOptions<M>> | FindOptions<M>
+export type QueryBeforeHook<M extends Model<any>, TContext = any> = (params: {
+  findOptions: FindOptionsWithAttributesWhere<M>
+  args: TArgs
+  context: TContext
+  info: TInfo
+}) =>
+  | FindOptionsWithAttributesWhere<M>
+  | Promise<FindOptionsWithAttributesWhere<M>>
 
-export type QueryAfterHook<M extends Model<any>> = (params: {
+export type QueryAfterHook<M extends Model<any>, TContext = any> = (params: {
   result: M | M[]
-  args: any
-  context: any
-  info: any
-}) => Promise<M | M[]> | M | M[]
+  args: TArgs
+  context: TContext
+  info: TInfo
+}) => M | M[] | Promise<M | M[]>
 
-export type CountAfterHook<M extends Model<any>> = (params: {
+export type CountAfterHook<M extends Model<any>, TContext = any> = (params: {
   result: number | GroupedCountResultItem[]
-  args: any
-  context: any
-  info: any
+  args: TArgs
+  context: TContext
+  info: TInfo
 }) =>
   | Promise<number | GroupedCountResultItem[]>
   | number
   | GroupedCountResultItem[]
 
-export type MutationBeforeHook<M extends Model<any>> = (params: {
+export type MutationBeforeHook<
+  M extends Model<any>,
+  TContext = any
+> = (params: {
   source: any
   args: TArgs
   context: TContext
   info: TInfo
 }) => any | Promise<any>
 
-export type MutationAfterHook<M extends Model<any>> = (params: {
-  newEntity: M
+export type MutationAfterHook<M extends Model<any>, TContext = any> = (params: {
+  createdEntity: M
   source: any
   args: TArgs
   context: TContext
@@ -139,16 +146,22 @@ export type MutationAfterHook<M extends Model<any>> = (params: {
   setWebhookData?: (callback: (defaultData: any) => any) => void
 }) => M | Promise<M>
 
-export type UpdateMutationAfterHook<M extends Model<any>> = (params: {
+export type UpdateMutationAfterHook<
+  M extends Model<any>,
+  TContext = any
+> = (params: {
   updatedEntity: M
-  entitySnapshot: M
+  previousPropertiesSnapshot: Partial<InferAttributes<M>>
   source: any
   args: TArgs
   context: TContext
   info: TInfo
 }) => M | Promise<M>
 
-export type DeleteMutationBeforeHook<M extends Model<any>> = (params: {
+export type DeleteMutationBeforeHook<
+  M extends Model<any>,
+  TContext = any
+> = (params: {
   where: Filterable<M>
   source: any
   args: TArgs
@@ -156,7 +169,10 @@ export type DeleteMutationBeforeHook<M extends Model<any>> = (params: {
   info: TInfo
 }) => Filterable<M> | Promise<Filterable<M>>
 
-export type DeleteMutationAfterHook<M extends Model<any>> = (params: {
+export type DeleteMutationAfterHook<
+  M extends Model<any>,
+  TContext = any
+> = (params: {
   deletedEntity: M
   source: any
   args: TArgs
@@ -164,11 +180,12 @@ export type DeleteMutationAfterHook<M extends Model<any>> = (params: {
   info: TInfo
 }) => M | Promise<M>
 
-export type SubscriptionFilterHook = (params: {
-  payload: any
-  args: TArgs
-  context: TContext
-}) => boolean | Promise<boolean>
+export type SubscriptionFilterHook<TContext = any> = (
+  payload: any,
+  args: TArgs,
+  context: TContext,
+  info: TInfo
+) => boolean | Promise<boolean>
 
 export type SubscriptionFilter = (params: {
   payload: any
@@ -176,67 +193,76 @@ export type SubscriptionFilter = (params: {
   context: any
 }) => boolean
 
-export type GraphqlSchemaDeclarationType = {
-  [key: string]: ModelDeclarationType<any> | GraphQLFieldConfig<any, any, any>
-}
-
-export type CreateFieldDeclarationType<M extends Model<any>> = {
-  before?: CreateBeforeHook<M> | CreateBeforeHook<M>[]
-  after?: CreateAfterHook<M> | CreateAfterHook<M>[]
+export type CreateFieldDeclarationType<M extends Model<any>, TContext = any> = {
+  before?: CreateBeforeHook<M, TContext> | CreateBeforeHook<M, TContext>[]
+  after?: CreateAfterHook<M, TContext> | CreateAfterHook<M, TContext>[]
   extraArg?: Record<string, { type: GraphQLInputType }>
+  subscriptionFilter?: SubscriptionFilterHook<TContext>
   preventDuplicateOnAttributes?: string[]
-  subscriptionFilter?: SubscriptionFilterHook
 }
 
-export type UpdateFieldDeclarationType<M extends Model<any>> = {
-  before?: UpdateBeforeHook<M> | UpdateBeforeHook<M>[]
-  after?: UpdateAfterHook<M> | UpdateAfterHook<M>[]
+export type GraphqlSchemaDeclarationType<TContext = any> = {
+  [key: string]:
+    | ModelDeclarationType<any, TContext>
+    | GraphQLFieldConfig<any, any, any>
+}
+
+export type UpdateFieldDeclarationType<M extends Model<any>, TContext = any> = {
+  before?: UpdateBeforeHook<M, TContext> | UpdateBeforeHook<M, TContext>[]
+  after?: UpdateAfterHook<M, TContext> | UpdateAfterHook<M, TContext>[]
   extraArg?: Record<string, { type: GraphQLInputType }>
-  subscriptionFilter?: SubscriptionFilterHook
+  subscriptionFilter?: SubscriptionFilterHook<TContext>
 }
 
-export type DeleteFieldDeclarationType<M extends Model<any>> = {
-  before?: DeleteBeforeHook<M> | DeleteBeforeHook<M>[]
-  after?: DeleteAfterHook<M> | DeleteAfterHook<M>[]
+export type DeleteFieldDeclarationType<M extends Model<any>, TContext = any> = {
+  before?: DeleteBeforeHook<M, TContext> | DeleteBeforeHook<M, TContext>[]
+  after?: DeleteAfterHook<M, TContext> | DeleteAfterHook<M, TContext>[]
   extraArg?: Record<string, { type: GraphQLInputType }>
-  subscriptionFilter?: SubscriptionFilterHook
+  subscriptionFilter?: SubscriptionFilterHook<TContext>
 }
 
-export type ListDeclarationType<M extends Model<any>> = {
+export type ListDeclarationType<M extends Model<any>, TContext = any> = {
   removeUnusedAttributes?: boolean
   enforceMaxLimit?: number
-  before?: QueryBeforeHook<M> | QueryBeforeHook<M>[]
-  after?: QueryAfterHook<M> | QueryAfterHook<M>[]
-  resolver?: Resolver
-  subscriptionFilter?: SubscriptionFilterHook
+  before?: QueryBeforeHook<M, TContext> | QueryBeforeHook<M, TContext>[]
+  after?: QueryAfterHook<M, TContext> | QueryAfterHook<M, TContext>[]
+  resolver?: GraphQLFieldResolver<any, TContext>
   contextToOptions?: boolean
   extraArg?: Record<string, { type: GraphQLInputType }>
   disableOptimizationForLimitOffset?: boolean
+  subscriptionFilter?: SubscriptionFilterHook<TContext>
 }
 
-export type ModelDeclarationType<M extends Model<any>> = {
+export type ModelDeclarationType<M extends Model<any>, TContext = any> = {
   model: ModelStatic<M>
   actions?: Action[]
   subscriptions?: Event[]
   webhooks?: WebhookType[]
-  before?: GlobalBeforeHook | GlobalBeforeHook[]
+  before?: GlobalBeforeHook<TContext> | GlobalBeforeHook<TContext>[]
   count?: {
     extraArg?: Record<string, { type: GraphQLInputType }>
-    before?: QueryBeforeHook<M> | QueryBeforeHook<M>[]
-    after?: CountAfterHook<M> | CountAfterHook<M>[]
-    resolver?: Resolver
+    before?: QueryBeforeHook<M, TContext> | QueryBeforeHook<M, TContext>[]
+    after?: CountAfterHook<M, TContext> | CountAfterHook<M, TContext>[]
+    resolver?: GraphQLFieldResolver<any, TContext>
   }
-  list?: ListDeclarationType<M>
-  create?: CreateFieldDeclarationType<M>
-  update?: UpdateFieldDeclarationType<M>
-  delete?: DeleteFieldDeclarationType<M>
+  list?: ListDeclarationType<M, TContext>
+  create?:
+    | CreateFieldDeclarationType<M, TContext>
+    | GraphQLFieldConfig<TSource, TContext, TArgs>
+  update?:
+    | UpdateFieldDeclarationType<M, TContext>
+    | GraphQLFieldConfig<TSource, TContext, TArgs>
+  delete?:
+    | DeleteFieldDeclarationType<M, TContext>
+    | GraphQLFieldConfig<TSource, TContext, TArgs>
   excludeFields?: string[]
   excludeFromRoot?: boolean
   type?: GraphQLObjectType
-  resolve?: Resolver
+  resolve?: GraphQLFieldResolver<any, TContext>
   args?: Record<string, { type: GraphQLInputType }>
   contextToOptions?: boolean
-  additionalSubscriptions?: SubscriptionList
+  additionalSubscriptions?: SubscriptionList<TContext>
+  additionalMutations?: MutationList<TContext>
 }
 
 export type SequelizeModels = { [key: string]: ModelStatic<any> } & {
@@ -247,8 +273,11 @@ export type GlobalPreCallback = (name: string) => Function | null
 
 export type Associations = { [key: string]: Association }
 
-export type GraphqlSequelizeResolverConfigType<M extends Model> = {
-  before?: QueryBeforeHook<M>
+export type GraphqlSequelizeResolverConfigType<
+  M extends Model,
+  TContext = any
+> = {
+  before?: QueryBeforeHook<M, TContext>
 }
 
 export interface GraphqlSequelizeResolverType {
@@ -265,50 +294,65 @@ export type Resolver = (
   info: any
 ) => any | Promise<any>
 
-export type CreateBeforeHook<M extends Model<any>> = (params: {
+export type CreateBeforeHook<M extends Model<any>, TContext = any> = (params: {
   source: any
   args: any
-  context: any
+  context: TContext
   info: any
 }) => Promise<any> | any
 
-export type CreateAfterHook<M extends Model<any>> = (params: {
-  newEntity: M
+export type CreateAfterHook<M extends Model<any>, TContext = any> = (params: {
+  createdEntity: M
   source: any
   args: any
-  context: any
+  context: TContext
   info: any
   setWebhookData: (callback: (defaultData: any) => any) => void
 }) => Promise<M> | M
 
-export type UpdateBeforeHook<M extends Model<any>> = (params: {
+export type UpdateBeforeHook<M extends Model<any>, TContext = any> = (params: {
   source: any
   args: any
-  context: any
+  context: TContext
   info: any
 }) => Promise<any> | any
 
-export type UpdateAfterHook<M extends Model<any>> = (params: {
+export type UpdateAfterHook<M extends Model<any>, TContext = any> = (params: {
   updatedEntity: M
-  entitySnapshot: any
+  previousPropertiesSnapshot: Partial<InferAttributes<M>>
   source: any
   args: any
-  context: any
+  context: TContext
   info: any
 }) => Promise<M> | M
 
-export type DeleteBeforeHook<M extends Model<any>> = (params: {
+export type DeleteBeforeHook<M extends Model<any>, TContext = any> = (params: {
   where: Filterable<M>
   source: any
   args: any
-  context: any
+  context: TContext
   info: any
 }) => Promise<Filterable<M>> | Filterable<M>
 
-export type DeleteAfterHook<M extends Model<any>> = (params: {
+export type DeleteAfterHook<M extends Model<any>, TContext = any> = (params: {
   deletedEntity: M
   source: any
   args: any
-  context: any
+  context: TContext
   info: any
 }) => Promise<void> | void
+
+// This type is made to ease the developer experience when using a GSG Hook.
+// Having less configuration scenarios reduce the potential sources of bugs and improve readability.
+export type FindOptionsWithAttributesWhere<M extends Model<any>> =
+  FindOptions<any> & {
+    // In the case of GraphQL, the where is an object with attributes
+    // as functions can only be applied on attributes based on their alias.
+    // Like { '$or' : {a: 1, b: 2} }
+    where: WhereAttributeHash<M>
+    // This should normally be either an Includeable or an array of Includeable[]
+    // We force it as an array of Includeable[] to make it easier to work with.
+    // This allow us to always initialize it and be able to use findOptions.include.push()
+    // without having to check if it's an array or not.
+    include: Includeable[]
+  }